requirements.txt
Django==4.0.6
psycopg2-binary==2.9.3
uwsgi==2.0.20
invoke==1.7.1
djangorestframework==3.13.1

celery==5.2.7